split sql and in memory engines
