+++ web/logs/constants.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+VALID_NICKS = {
+    'Cosmo': ['cosmo', 'cfumo'],
+    'Graham': ['graham', 'jorgon'],
+    'Jesse': ['jesse', 'je-c'],
+    'Will': ['will', 'wyll', 'wyll_'],
+    'Zhenya': ['zhenyah', 'zhenya', 'zdog', 'swphantom', 'za', 'zhenya2'],
+}
